+++ docs/html/search/files_0.js
@@ -3,5 +3,6 @@ var searchData=
   ['chip_5fid_2eh_0',['chip_id.h',['../chip__id_8h.html',1,'']]],
   ['counter0_5fmode0_2eh_1',['counter0_mode0.h',['../counter0__mode0_8h.html',1,'']]],
   ['counter0_5fmode1_2eh_2',['counter0_mode1.h',['../counter0__mode1_8h.html',1,'']]],
-  ['counter2_5fmode0_2eh_3',['counter2_mode0.h',['../counter2__mode0_8h.html',1,'']]]
+  ['counter0_5fmode1_5fexample_2ec_3',['counter0_mode1_example.c',['../counter0__mode1__example_8c.html',1,'']]],
+  ['counter2_5fmode0_2eh_4',['counter2_mode0.h',['../counter2__mode0_8h.html',1,'']]]
 ];